migrate to the DBConnection framework
The main motivation of this migration is to maintain a pool of open connections and
keep those connections alive. So, when a query gets executed, the open connection is
received from the pool, thus we don't spend time on opening a new connection.

This migration brought a few changes to the api.

The query method requires the pid of the pool process or the pool_name.

    {:ok, pid} = Bolt.Sips.start_link([])

    Bolt.Sips.query!(pid, query)
    Bolt.Sips.query!(Bolt.Sips.pool_name, query)

The Bolt.Sips.rollback/1 and Bolt.Sips.commit/1 release an open connection,
so the connection is returned to the pool and cannot be used anymore:

   conn = Bolt.Sips.begin(pid)
   Bolt.Sips.query!(conn, query)
   Bolt.Sips.commit(conn)
   Bolt.Sips.query!(conn, another_query) # raises an error

The Bolt.Sips.Protocol.handle_execute/4 method tries to use an open connection
which is stored in the state of a process. If the connection is broken,
it opens connection, executes a query and stores the new connection as
the state, thus, it can be reused by other queries.

It would be perfect to ping each N secs a Neo4j server in order to verify
open connections, but Neo4j doesn't provide any message
(https://boltprotocol.org/v1/#messages) which can be used for that.
The doc of the bolt protocol says that connections are closed by clients.
So, Neo4j shouldn't close connections, but Neo4j instances offered by
https://www.graphenedb.com do close connections in a few hours.
At least it is true for the hobby plan.